DECRYPTION/requirement.txt
asgiref==3.3.4
azure-common==1.1.27
azure-core==1.14.0
azure-identity==1.6.0
azure-keyvault-keys==4.4.0
azure-keyvault-secrets==4.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.1.11
idna==2.10
install==1.3.4
isodate==0.6.0
MarkupSafe==2.0.1
msal==1.12.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.0
Pillow==7.2.0
portalocker==1.7.1
pycparser==2.20
pycryptodome==3.9.9
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
pywin32==300
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
django-python3-ldap==0.11.4
gunicorn==20.1.0